requirements.txt
fastapi
uvicorn
sqlalchemy
sqlalchemy-utils 
databases
mysql-connector-python
python-dotenv